backend/requirements.txt
# ─── Web Framework ────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0

# ─── Database ─────────────────────────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1

# ─── Validation / Settings ────────────────────────────────────────────────────
pydantic[email]==2.7.1
pydantic-settings==2.2.1

# ─── Auth ─────────────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
python-multipart==0.0.9

# ─── Cache / Queue ────────────────────────────────────────────────────────────
redis==5.0.4
celery==5.4.0

# ─── AI ───────────────────────────────────────────────────────────────────────
openai==1.30.1

# ─── File Processing ──────────────────────────────────────────────────────────
pdfplumber==0.11.0
python-magic==0.4.27

# ─── Rate Limiting ────────────────────────────────────────────────────────────
slowapi==0.1.9

# ─── Observability ────────────────────────────────────────────────────────────
opentelemetry-sdk==1.24.0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-instrumentation-sqlalchemy==0.45b0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
prometheus-client==0.20.0

# ─── Logging ──────────────────────────────────────────────────────────────────
structlog==24.2.0

# ─── DB sync driver (Alembic offline migrations + psycopg2 fallback) ──────────
psycopg2-binary==2.9.9

# ─── Testing ──────────────────────────────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.7
pytest-cov==5.0.0
httpx==0.27.0
aiosqlite==0.20.0
